refactor and add translated validation messages
